+++ CHANGELOG.md
@@ -1,2 +1,16 @@
-# WIP
+# Change Log
+
+### version: 1.0.1
+
+- Added `CelikApi.disconnectCard()`
+- Added posiibility for custom implementation, using `CelikDataAPI` and `CelikAPDUBase`
+- Better file structure
+- Improved comments
+- Fixed library name
+- Fix: now running PCSC in `newIsolate = true`
+- Simple test added
+- Updated `README.md`
+
 ### version: 1.0.0+1
+
+Test version